Fix buffer overflow

+++ args.h
@@ -1,7 +1,7 @@
 #ifndef _ARGS_H
 #define _ARGS_H
 
-int get_args(int argc, char* argv[], char* space, char* alpha);
+size_t get_args(int argc, char* argv[], char* space, char* alpha);
 void print_version(char* program_name);
 void print_help(char* program_name);
